+++ src/App.js
@@ -5,7 +5,6 @@ import Navbar from "./components/Navbar";
 import Dashboard from "./components/Dashboard";
 import Form from "./components/Form";
 import Favorites from "./components/Favourites";
-import Favourites2 from "./components/Favourites2";
 import Search from "./components/Search";
 import Appbar from "./components/Appbar";
 import 'semantic-ui-css/semantic.min.css'
@@ -32,7 +31,7 @@ function App() {
                         <Route exact strict path="/login" component={Login}/>
                         <Route exact strict path="/user" component={LoggedUser}/>
                         <Route exact strict path="/form/" component={Form}/>
-                        <Route exact strict path="/favs" component={Favourites2}/>
+                        <Route exact strict path="/favs" component={Favorites}/>
                         <Route component={() => <h1>404 - sadface</h1>}/>
                     </Switch>
                 </main>
